odoo_nw_production/print_server/requirements.txt
flask==3.0.0
python-escpos==3.0
PyPDF2==3.0.1
pycups==2.0.1
pyusb==1.2.1
